+++ source/tests/end2end/imaginary_double1/lib.kefir.c
@@ -29,22 +29,22 @@ int f64_compat[] = {
     CLASSIFY(((_Imaginary double) 1.0if) - (_Imaginary double) 2.0if),
     CLASSIFY(((_Imaginary double) 1.0if) * (_Imaginary double) 2.0if),
     CLASSIFY(((_Imaginary double) 1.0if) / (_Imaginary double) 2.0if),
-    CLASSIFY(((float) 1.0if) + (_Imaginary double) 2.0if),
-    CLASSIFY(((float) 1.0if) - (_Imaginary double) 2.0if),
-    CLASSIFY(((float) 1.0if) * (_Imaginary double) 2.0if),
-    CLASSIFY(((float) 1.0if) / (_Imaginary double) 2.0if),
-    CLASSIFY(((_Imaginary double) 1.0if) + (float) 2.0if),
-    CLASSIFY(((_Imaginary double) 1.0if) - (float) 2.0if),
-    CLASSIFY(((_Imaginary double) 1.0if) * (float) 2.0if),
-    CLASSIFY(((_Imaginary double) 1.0if) / (float) 2.0if),
-    CLASSIFY(((_Complex float) 1.0if) + (_Imaginary double) 2.0if),
-    CLASSIFY(((_Complex float) 1.0if) - (_Imaginary double) 2.0if),
-    CLASSIFY(((_Complex float) 1.0if) * (_Imaginary double) 2.0if),
-    CLASSIFY(((_Complex float) 1.0if) / (_Imaginary double) 2.0if),
-    CLASSIFY(((_Imaginary double) 1.0if) + (_Complex float) 2.0if),
-    CLASSIFY(((_Imaginary double) 1.0if) - (_Complex float) 2.0if),
-    CLASSIFY(((_Imaginary double) 1.0if) * (_Complex float) 2.0if),
-    CLASSIFY(((_Imaginary double) 1.0if) / (_Complex float) 2.0if)
+    CLASSIFY(((double) 1.0if) + (_Imaginary double) 2.0if),
+    CLASSIFY(((double) 1.0if) - (_Imaginary double) 2.0if),
+    CLASSIFY(((double) 1.0if) * (_Imaginary double) 2.0if),
+    CLASSIFY(((double) 1.0if) / (_Imaginary double) 2.0if),
+    CLASSIFY(((_Imaginary double) 1.0if) + (double) 2.0if),
+    CLASSIFY(((_Imaginary double) 1.0if) - (double) 2.0if),
+    CLASSIFY(((_Imaginary double) 1.0if) * (double) 2.0if),
+    CLASSIFY(((_Imaginary double) 1.0if) / (double) 2.0if),
+    CLASSIFY(((_Complex double) 1.0if) + (_Imaginary double) 2.0if),
+    CLASSIFY(((_Complex double) 1.0if) - (_Imaginary double) 2.0if),
+    CLASSIFY(((_Complex double) 1.0if) * (_Imaginary double) 2.0if),
+    CLASSIFY(((_Complex double) 1.0if) / (_Imaginary double) 2.0if),
+    CLASSIFY(((_Imaginary double) 1.0if) + (_Complex double) 2.0if),
+    CLASSIFY(((_Imaginary double) 1.0if) - (_Complex double) 2.0if),
+    CLASSIFY(((_Imaginary double) 1.0if) * (_Complex double) 2.0if),
+    CLASSIFY(((_Imaginary double) 1.0if) / (_Complex double) 2.0if)
 };
 
 _Imaginary double f64_arr[] = {
@@ -86,40 +86,40 @@ _Imaginary double f64_arr[] = {
 
 _Imaginary double f64_ops[] = {
     ((_Imaginary double) 3.0if) + ((_Imaginary double) 9.0if),
-    ((_Imaginary double) 3.0if) + ((float) 9.0f),
-    ((float) 3.0f) + ((_Imaginary double) 9.0if),
-    ((_Imaginary double) 3.0if) + ((_Complex float) 6.0f + 9.0if),
-    ((_Complex float) 6.0f + 9.0if) + ((_Imaginary double) -3.0if),
+    ((_Imaginary double) 3.0if) + ((double) 9.0f),
+    ((double) 3.0f) + ((_Imaginary double) 9.0if),
+    ((_Imaginary double) 3.0if) + ((_Complex double) 6.0f + 9.0if),
+    ((_Complex double) 6.0f + 9.0if) + ((_Imaginary double) -3.0if),
     ((_Imaginary double) 3.1if) + (long) -5000,
     ((_Imaginary double) 3.2if) + (unsigned long) 5000,
     ((long) -5000) + ((_Imaginary double) 3.4if),
     ((unsigned long) 5000) + ((_Imaginary double) 3.5if),
 
     ((_Imaginary double) 3.0if) - ((_Imaginary double) 9.0if),
-    ((_Imaginary double) 3.0if) - ((float) 9.0f),
-    ((float) 3.0f) - ((_Imaginary double) 9.0if),
-    ((_Imaginary double) 3.0if) - ((_Complex float) 6.0f + 9.0if),
-    ((_Complex float) 6.0f + 9.0if) - ((_Imaginary double) -3.0if),
+    ((_Imaginary double) 3.0if) - ((double) 9.0f),
+    ((double) 3.0f) - ((_Imaginary double) 9.0if),
+    ((_Imaginary double) 3.0if) - ((_Complex double) 6.0f + 9.0if),
+    ((_Complex double) 6.0f + 9.0if) - ((_Imaginary double) -3.0if),
     ((_Imaginary double) 3.1if) - (long) -5000,
     ((_Imaginary double) 3.2if) - (unsigned long) 5000,
     ((long) -5000) - ((_Imaginary double) 3.4if),
     ((unsigned long) 5000) - ((_Imaginary double) 3.5if),
 
     ((_Imaginary double) 3.0if) * ((_Imaginary double) 9.0if),
-    ((_Imaginary double) 3.0if) * ((float) 9.0f),
-    ((float) 3.0f) * ((_Imaginary double) 9.0if),
-    ((_Imaginary double) 3.0if) * ((_Complex float) 6.0f + 9.0if),
-    ((_Complex float) 6.0f + 9.0if) * ((_Imaginary double) -3.0if),
+    ((_Imaginary double) 3.0if) * ((double) 9.0f),
+    ((double) 3.0f) * ((_Imaginary double) 9.0if),
+    ((_Imaginary double) 3.0if) * ((_Complex double) 6.0f + 9.0if),
+    ((_Complex double) 6.0f + 9.0if) * ((_Imaginary double) -3.0if),
     ((_Imaginary double) 3.1if) * (long) -5000,
     ((_Imaginary double) 3.2if) * (unsigned long) 5000,
     ((long) -5000) * ((_Imaginary double) 3.4if),
     ((unsigned long) 5000) * ((_Imaginary double) 3.5if),
 
     ((_Imaginary double) 3.0if) / ((_Imaginary double) 9.0if),
-    ((_Imaginary double) 3.0if) / ((float) 9.0f),
-    ((float) 3.0f) / ((_Imaginary double) 9.0if),
-    ((_Imaginary double) 3.0if) / ((_Complex float) 6.0f + 9.0if),
-    ((_Complex float) 6.0f + 9.0if) / ((_Imaginary double) -3.0if),
+    ((_Imaginary double) 3.0if) / ((double) 9.0f),
+    ((double) 3.0f) / ((_Imaginary double) 9.0if),
+    ((_Imaginary double) 3.0if) / ((_Complex double) 6.0f + 9.0if),
+    ((_Complex double) 6.0f + 9.0if) / ((_Imaginary double) -3.0if),
     ((_Imaginary double) 3.1if) / (long) -5000,
     ((_Imaginary double) 3.2if) / (unsigned long) 5000,
     ((long) -5000) / ((_Imaginary double) 3.4if),
@@ -131,40 +131,40 @@ _Imaginary double f64_ops[] = {
 
 double f64_ops2[] = {
     ((_Imaginary double) 3.0if) + ((_Imaginary double) 9.0if),
-    ((_Imaginary double) 3.0if) + ((float) 9.0f),
-    ((float) 3.0f) + ((_Imaginary double) 9.0if),
-    ((_Imaginary double) 3.0if) + ((_Complex float) 6.0f + 9.0if),
-    ((_Complex float) 6.0f + 9.0if) + ((_Imaginary double) -3.0if),
+    ((_Imaginary double) 3.0if) + ((double) 9.0f),
+    ((double) 3.0f) + ((_Imaginary double) 9.0if),
+    ((_Imaginary double) 3.0if) + ((_Complex double) 6.0f + 9.0if),
+    ((_Complex double) 6.0f + 9.0if) + ((_Imaginary double) -3.0if),
     ((_Imaginary double) 3.1if) + (long) -5000,
     ((_Imaginary double) 3.2if) + (unsigned long) 5000,
     ((long) -5000) + ((_Imaginary double) 3.4if),
     ((unsigned long) 5000) + ((_Imaginary double) 3.5if),
 
     ((_Imaginary double) 3.0if) - ((_Imaginary double) 9.0if),
-    ((_Imaginary double) 3.0if) - ((float) 9.0f),
-    ((float) 3.0f) - ((_Imaginary double) 9.0if),
-    ((_Imaginary double) 3.0if) - ((_Complex float) 6.0f + 9.0if),
-    ((_Complex float) 6.0f + 9.0if) - ((_Imaginary double) -3.0if),
+    ((_Imaginary double) 3.0if) - ((double) 9.0f),
+    ((double) 3.0f) - ((_Imaginary double) 9.0if),
+    ((_Imaginary double) 3.0if) - ((_Complex double) 6.0f + 9.0if),
+    ((_Complex double) 6.0f + 9.0if) - ((_Imaginary double) -3.0if),
     ((_Imaginary double) 3.1if) - (long) -5000,
     ((_Imaginary double) 3.2if) - (unsigned long) 5000,
     ((long) -5000) - ((_Imaginary double) 3.4if),
     ((unsigned long) 5000) - ((_Imaginary double) 3.5if),
 
     ((_Imaginary double) 3.0if) * ((_Imaginary double) 9.0if),
-    ((_Imaginary double) 3.0if) * ((float) 9.0f),
-    ((float) 3.0f) * ((_Imaginary double) 9.0if),
-    ((_Imaginary double) 3.0if) * ((_Complex float) 6.0f + 9.0if),
-    ((_Complex float) 6.0f + 9.0if) * ((_Imaginary double) -3.0if),
+    ((_Imaginary double) 3.0if) * ((double) 9.0f),
+    ((double) 3.0f) * ((_Imaginary double) 9.0if),
+    ((_Imaginary double) 3.0if) * ((_Complex double) 6.0f + 9.0if),
+    ((_Complex double) 6.0f + 9.0if) * ((_Imaginary double) -3.0if),
     ((_Imaginary double) 3.1if) * (long) -5000,
     ((_Imaginary double) 3.2if) * (unsigned long) 5000,
     ((long) -5000) * ((_Imaginary double) 3.4if),
     ((unsigned long) 5000) * ((_Imaginary double) 3.5if),
 
     ((_Imaginary double) 3.0if) / ((_Imaginary double) 9.0if),
-    ((_Imaginary double) 3.0if) / ((float) 9.0f),
-    ((float) 3.0f) / ((_Imaginary double) 9.0if),
-    ((_Imaginary double) 3.0if) / ((_Complex float) 6.0f + 9.0if),
-    ((_Complex float) 6.0f + 9.0if) / ((_Imaginary double) -3.0if),
+    ((_Imaginary double) 3.0if) / ((double) 9.0f),
+    ((double) 3.0f) / ((_Imaginary double) 9.0if),
+    ((_Imaginary double) 3.0if) / ((_Complex double) 6.0f + 9.0if),
+    ((_Complex double) 6.0f + 9.0if) / ((_Imaginary double) -3.0if),
     ((_Imaginary double) 3.1if) / (long) -5000,
     ((_Imaginary double) 3.2if) / (unsigned long) 5000,
     ((long) -5000) / ((_Imaginary double) 3.4if),
@@ -258,11 +258,11 @@ double fi64_div(_Imaginary double x, _Imaginary double y) {
     return x / y;
 }
 
-_Complex float fi64_f64_add(_Imaginary double x, double y) {
+_Complex double fi64_f64_add(_Imaginary double x, double y) {
     return x + y;
 }
 
-_Complex float fi64_f64_sub(_Imaginary double x, double y) {
+_Complex double fi64_f64_sub(_Imaginary double x, double y) {
     return x - y;
 }
 
@@ -290,34 +290,34 @@ _Imaginary double f64_fi64_div(double x, _Imaginary double y) {
     return x / y;
 }
 
-_Complex float cf64_fi64_add(_Complex double x, _Imaginary double y) {
+_Complex double cf64_fi64_add(_Complex double x, _Imaginary double y) {
     return x + y;
 }
 
-_Complex float cf64_fi64_sub(_Complex double x, _Imaginary double y) {
+_Complex double cf64_fi64_sub(_Complex double x, _Imaginary double y) {
     return x - y;
 }
 
-_Complex float cf64_fi64_mul(_Complex double x, _Imaginary double y) {
+_Complex double cf64_fi64_mul(_Complex double x, _Imaginary double y) {
     return x * y;
 }
 
-_Complex float cf64_fi64_div(_Complex double x, _Imaginary double y) {
+_Complex double cf64_fi64_div(_Complex double x, _Imaginary double y) {
     return x / y;
 }
 
-_Complex float fi64_cf64_add(_Imaginary double x, _Complex double y) {
+_Complex double fi64_cf64_add(_Imaginary double x, _Complex double y) {
     return x + y;
 }
 
-_Complex float fi64_cf64_sub(_Imaginary double x, _Complex double y) {
+_Complex double fi64_cf64_sub(_Imaginary double x, _Complex double y) {
     return x - y;
 }
 
-_Complex float fi64_cf64_mul(_Imaginary double x, _Complex double y) {
+_Complex double fi64_cf64_mul(_Imaginary double x, _Complex double y) {
     return x * y;
 }
 
-_Complex float fi64_cf64_div(_Imaginary double x, _Complex double y) {
+_Complex double fi64_cf64_div(_Imaginary double x, _Complex double y) {
     return x / y;
 }
